fix merge errors
