+++ src/pages/events/[uuid].astro
@@ -1,6 +1,6 @@
 ---
 import Layout from '../../layouts/Layout.astro';
-import Challenge from '../../components/ChallengesOld.svelte';
+import Challenge from '../../components/Challenges.svelte';
 import Header from '../../components/Header.svelte';
 import { normalWrapper } from '../../lib/db';
